# backend/openapi/palmchain-v2.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PalmChain V2 API",
    "version": "2.0.0-phase10",
    "description": "Security-first Phase 2-10 contract with fail-closed pilot readiness and privacy-minimized operational evidence."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": {
      "oidc": { "type": "openIdConnect", "openIdConnectUrl": "https://identity-owner.example/.well-known/openid-configuration" }
    },
...
  },
  "security": [{ "oidc": [] }],
  "paths": {
...
    "/consent": {
      "post": { "operationId": "grantConsent", "responses": { "200": { "description": "Granular, purpose-bound consent" }, "400": { "description": "Invalid or over-broad consent" } } }
    }
  }
}